requirements.txt
kafka-python==2.0.2
Faker==4.1.0